Update to kind-0.11.1. Override mtu manually. (#23)
* Update to kind-0.11.1. Override mtu manually.

* Inject a systemd unit file for docker0 mtu.

The MTU on the docker0 nic is not set by the daemon.json
file, but it appears to be needed. We add a systemd service
to be run after the dockerd starts to adjust it. This should
be very robust against updates and also reboot-safe.

* Write_files from files input instead of b64.

* Make docker MTU configurable (default 1400).

Variable gets injected via cloud-init. Avoid external files.

* Change to canonical tf formatting.

terraform fmt helped.

Signed-off-by: Kurt Garloff <scs@garloff.de>