Finally working app credentials: project_id needed.
Well, with application credential, we don't need the project_id.
And openstackclient does actively dislike it, aksing for a project
scoped token, which does not work.
cluster-api-provider-openstack on the other hand NEEDS it, as it
does not consider Authinfo without projectID to be valid. It knows
how to get projectID from projectName, but not how to get it from
applicationCredential. (Maybe gophercloud does not have the needed
helpers.)

Signed-off-by: Kurt Garloff <kurt@garloff.de>