Feat/appcred (#51)
* Create and use application credential (#47).

This avoids leaking the user password into the k8s cluster management.

* Fix object reference. Add cleanup for appcred.

* Explicitly set auth_type. Avoid passing unneeded IDs.

* Make tf fmt happy.

* Add OS_CLOUD to environment, bash convenience.

* Document the usage of application credentials.

* Patch installed keystoneauth1 to fix bug 1876317.

* Finally working app credentials: project_id needed.

Well, with application credential, we don't need the project_id.
And openstackclient does actively dislike it, asking for a project
scoped token, which does not work.
cluster-api-provider-openstack on the other hand NEEDS it, as it
does not consider AuthInfo without ProjectID to be valid. It knows
how to get ProjectID from ProjectName, but not how to get it from
ApplicationCredential. (Maybe gophercloud does not have the needed
helpers.)

Signed-off-by: Kurt Garloff <kurt@garloff.de>