Tune etcd heartbeat and use SSD flavor for control nodes.
We see too many etcd "too slow" and timeouts and reelections
in some test setups. etcd is sensitive to slow disk writes
(which are persisted by fsync) -- so disks which are shared
and may have latencies due to queued I/O operatioms from
other processes and/or hosts.

We address this in two ways:
* We increase the heartbeat-interval (and the corresponding
  elevtion-timeout) from 100ms to 250ms (and 1s to 2.5s resp.).
  This makes us more tolerant against slow writes and roughly
  halves the number of etcd failovers on shared ceph disks.
* We change the default flavor for the controller nodes to
  -20s (a local 20GB SSD) for storage.

Signed-off-by: Kurt Garloff <kurt@garloff.de>