Stop patching etcd.yaml in uncoordinated ways.
Trouble is that the restart causes us to lose quorum if there is no
other etcd running (yet). So restarting would need to wait for others
and we have no simple way to ensure it.

So let's disable this for now. Best would be to have the default config
with the changed heartbeat/election intervals.

Signed-off-by: Kurt Garloff <kurt@garloff.de>