Delete pods and pvcs when deleting cluster.
We want to avoid leaving OpenStack resources behind.
Thus we want to clean up persistent volume claims.
To be able to do so, these must not be in-use, which we achieve
by also deleting all pods.

To be done: We probably would need to treat loadbalancers like
persistent volume claims. (Will do after we have good test cases
for these.)

Signed-off-by: Kurt Garloff <kurt@garloff.de>