Feat/mtu detection (#110)
* Add autodetection for MTU. (#108)

This is achieved by injecting get_mtu.sh.
If kind_mtu is set to 0 (new default), then the cloud's MTU will
be taken from the NIC with the default route. It will be forwarded
to docker/kind. The k8s cluster gets 50 bytes smaller MTU.
If an MTU is set, it will be used for docker/kind; k8s cluster
will again be 50b smaller.

Sizes are rounded down to full 8 bytes.

* Remove spurious comma

Signed-off-by: Kurt Garloff <kurt@garloff.de>