Feat/per cluster k8s image (#128)
* Move wait_capi_image.sh to create_cluster.sh.

The rationale is that the capi image is per-cluster and not global.
So we don't wait for a capi image in the cluster initialization.
Instead, it's done from create_cluster.sh, where we parse the
clusterctl-${CLUSTERNAME}.yaml to determine which image we need to
have available. If it's missing, we call upload_capi_image.sh and
grab and register the image.
We still start wait_capi_image.sh early on in bootstrap.sh (for the
specified default testcluster image), so we parallelize things.

Being at it: wait_capi_image.sh and upload_capi_image.sh are now
untemplated and get all their settings from clusterctl.yaml.
upload_capi_image.sh did some generic openstack setup, which is
now pushed to the new prepare_openstack.sh and called early on
from bootstrap.sh. We also install yq from bootstrap.sh at the
beginning.

Also really evaluate OPENSTACK_IMAGE_REGISTRATION_EXTRA_FLAGS.

* Add comments for TODOS: Project and AppCred creation.

In our nextgen implementation, we want to have AppCreds per cluster.
A pair of them, actually.
And optionally even an own project ...

Signed-off-by: Kurt Garloff <kurt@garloff.de>