Use kustomize to enable monitor (and proxy). (#175)
* Use kustomize to enable monitor (and proxy).

This is a different approach to fixing #173:
We can explicitly ask the OCCM to use a health monitor, see discussion
in #174. This is preferable, as we can leave the LB defaults as is.

Use kustomize to actually do the changes to the downloaded nginx
deployment file, as this is a rather robust way.

Being at it, also allow enabling the proxy protocol, so out backends
see the real IP (in the http header set by our nginx) and not just the
VIP from the LB.

We save the generated nginx deployment file, so we can use it for
cleanup in delete_cluster.sh again.

Also: Ship the nginx-i-c-v1.1.2 deployment file directly.

* Error out on failed download or kustomization.

* Use the right ConfigMap settings.

Follow
https://kubernetes.github.io/ingress-nginx/user-guide/nginx-configuration/configmap
not some possible outdated internet docs.

* For completeness: Add (unused) kustomization for nomonitor

Someone might decide that using a ClusterIP and thus not requiring
a health-monitor might be cleaner. I don't think it's a good idea,
but who knows ...

* Add release note snipped on monitor and proxy.

Signed-off-by: Kurt Garloff <kurt@garloff.de>