Only copy get_mtu/get_k8s_git/wait.sh into /tmp.
We have three scripts left that get copied over from the deployment
machine and executed early during the bootstrap. We moved them to /tmp,
so they don't clutter the user's home dir.

The git branch to be used is passed automatically now, it's passed
as a variable.
(We also have an option to propagate local changes, but this is not
yet implemented. For now only detect and warn if such changes exist.)

Signed-off-by: Kurt Garloff <kurt@garloff.de>