Feat/repo on mgmtcluster (#176)
* Use git checkout on management node.

We change the installation of scripts, templates on the management node
significantly:
* We do a git checkout from the k8s-cluster-api-provider repo on the
  management node. This makes it a lot easier to do incremental updates,
  as we don't need to redeploy the management host to update.
* We reduce the terraform templating mechanism, but instead copy
  settings into ~/.capi-settings, ~/cluster-defaults/cloud.conf,
  ~/cluster-defaults/clouds.yaml, ~/cluster-defaults/clusterctl.yaml
* When creating clusters, we copy the settings from ~/cluster-defaults/
  into ~/$CLUSTER_NAME/  (TODO: Partial copying ...)
* All yaml files that we apply to the cluster are saved in the
  ~/$CLUSTER_NAME/ directory. (TODO: Remove directory again in
  delete_cluster.sh)
* We have moved some common boilerplate to ~/bin/cccfg.inc

* Fix get-kubeconfig to work with new file locations.

* Fix mgmtcluster script deployments.

* Allow passing branch to check out on instance.

* Fix remaining wrong file locations.

* Make terraform fmt happy.

* Fix extracting cloud_provider from clusterctl.yaml.

* Only copy get_mtu/get_k8s_git/wait.sh into /tmp.

We have three scripts left that get copied over from the deployment
machine and executed early during the bootstrap. We moved them to /tmp,
so they don't clutter the user's home dir.

The git branch to be used is passed automatically now, it's passed
as a variable.
(We also have an option to propagate local changes, but this is not
yet implemented. For now only detect and warn if such changes exist.)

* Clean up after helm install, make ARCH tolerant.

Support non-x86-64 in script. Clean up after install.
Use latest helm-v3.8.1

* Prepare for unrestricted app-cred, public IP.

* Start documenting that clustercfg is in subdir now.

* We must not use quoted filenames with "~".

Main usage was to ensure we properly escape whitespace
in the $CLUSTER_NAME. Support was incomplete anyhow, so
let's rather disallow this and keep things simple.

* Fix filenames for get-kubeconfig and clean.

* Fill in missing files from cluster-defaults.

* Fix typo in metrics-server deployment.

* Move deployed manifests to subdirectory deployed-manifests.d/

* Also remember flux installation. Note on cleanup.

* Pass --os-cloud=$PROVIDER in delete_cluster.sh.

This may be called from ssh and we may not have the benefit of
OS_CLOUD having been set for us in ~/.bash_aliases.

* Add export OS_CLOUD to cccfg.inc, so it's always set correctly.

* Document new setup with git repo on mgmt node.

Signed-off-by: Kurt Garloff <kurt@garloff.de>