Disable the usage of proxy protocol by default.
If nginx expects inbound connections to use it, this works fine when the
connections come from the LB that has been configured to use the proxy
protocol. However, talking to ingress directly from the pods does then
no longer work (as noone injects the proxy proto then).

This means we should probably not enable it by default.

Signed-off-by: Kurt Garloff <kurt@garloff.de>