Feat/add kustpatch (#190)
* Add note on add_cluster-network.sh.

* Add script that creates dir structure exp. from kustomize.

If you just want too use kustomize to do context aware patching,
having all the directory infratructure and kustomization.yaml
files is an annoyance. Let's make this simpler by allowing
kustpatch.sh < input.yaml patch1.yaml [...] > output.yaml

* Strip trailing / from passed cluster name.

This is pure convenience -- user may use tab completion which
results in a trailing / for the cluster name (as we have a directory
with this name). Strip it ...

Signed-off-by: Kurt Garloff <kurt@garloff.de>