Latest OCCM templates use --cluster-name=${CLUSTER_NAME}.
So our patch to actually set the cluster name is no longer
required. Unfortuantely, the OCCM template sets kubernetes
as cluster-name, so the whole upstream improvement is useless.

We now are back to the issue that we have conflicting loadbalancer
names.

We need to use a different patch to set the CLUSTER_NAME ...
Detect new template and to this!

Signed-off-by: Kurt Garloff <kurt@garloff.de>