Support case where cloud_provider unequals ENVIRONMENT.
The generatio of mycloud.$(ENVIRONMENT).yaml assumed that the cloud has
the same name as ENVIRONMENT. This is not always true; users are free
to chose cloud_provider="SomeOtherName" in
environments/environment-${ENVIRONMENT}.tfvars.

We now correctly extract cloud_provider from it and extract the
${cloud_provider} info from clouds.yaml and put it into
mycloud.${cloud_provider}.yaml where terraform looks for it.

Should be included in 3.x and 3.0.x.

Signed-off-by: Kurt Garloff <kurt@garloff.de>