Make sure we wait longer for new clusters (#208)
There was logic to ensure we wait longer for new clusters, as the
control plane we wait does not even exist if we check too early,
so the script may error out.
Unfortunately, the logic was reversed, which is fixed now.

Should be backported to 3.x and 3.0.x after more testing.

Signed-off-by: Kurt Garloff <kurt@garloff.de>