Find ports that have been left over and delete them.
If a VM or a loadbalancer failed and had to be recreated, we may be
faced with remaining ports and these then prevent the security group
to be deleted which make kubectl wait forever for the cleanup.
Detect this and remove ports in delete_cluster.sh

This implements the approach described in #211.

Signed-off-by: Kurt Garloff <kurt@garloff.de>