Source ~/.bash_aliases from some scripts.
We need to do this when these scripts are invoked remotely (via ssh from
the Makefile), as non-interactive shells do not read ~/.bashrc.

Do this for cleanup.sh, deploy_cluster_api.sh and sonobuoy.sh which are
the only scripts called directly from the Makefile AFAICS.

This should address issue #240.

Signed-off-by: Kurt Garloff <kurt@garloff.de>