Enable etcd tweaks by default.
This keeps the etcd clusters stable even in somewhat awkward situations
with network storage (or high steal time, though that does not seem to
be as muh of a problem).

It's better to build stable 3/5/7 node clusters for the control plane
and risk losing etcd updates when a node really crashes than not having
stable clusters.

So let's default those tweak to be enabled.

Signed-off-by: Kurt Garloff <kurt@garloff.de>