Convert all CC-BY-SA-4.0 licensed files to Apache-2.0.
As we are closely connected to a bunch of Apache-2.0 licensed code,
chosing CC-BY-SA-4.0 would create a repository with mixed ASL-2.0
and CC-BY-SA-4.0 content; Apache-2.0 should be compatible, so that
the whole combined thing ends up being CC-BY-SA-4.0. However, the
devil is in the details and we don't want to feed the lawyers.

So simplify things by putting everything under Apache-2.0.

Signed-off-by: Kurt Garloff <kurt@garloff.de>