Change controller_flavor default to SCS-2C:4:20s
This means 2 dedicated cores, 4GiB RAM, 20GB (local) SSD storage.
It had a typo before, starting with SCS-2D.

Note that we have recently agreed on requiring a flavor SCS-2V:4:20s
in SCS-compliant IaaS offerings. So we will switch to this soon, as
defaulting to something that must exist is a better choice.

Signed-off-by: Kurt Garloff <kurt@garloff.de>