Fix needrestart configuration
Needrestart is part of the `apt-get upgrade` process from Ubuntu 22.04.
By default, it is set to an "interactive" mode which causes
the interruption of scripts. The interactive mode is applied when the new
kernel version is available after the upgrade process and when upgraded
services need to restart.

This commit adds a custom configuration file that overrides mentioned
settings and ensures that kernel hints are printed only to stderr and
services are restarted automatically if needed.

Signed-off-by: Matej Feder <matej.feder@dnation.cloud>