Fix deletion of management server floating IP in cleanup script
The cleanup script's function `resourcelist` wrongly filtered the floating IP of
the management server. The function returned brackets `]}` on the end of the
floating IP string which causes an issue.

This commit extends `resourcelist` function arguments with an argument that allows
a custom definition of `sed` expression which filters the output of the function.
The custom `sed` expression is then used to filter the management server floating
IP without unwanted brackets on the end.

Issue: #342

Signed-off-by: Matej Feder <matej.feder@dnation.cloud>