Fix configure_containerd.sh script idempotency (#472)
* Fix configure_containerd.sh script idempotency

Script configure_containerd.sh did not evaluate whether the
containerd file is already present in the cluster-template.yaml.
Hence, the file could be added on each script execution.

This commit adds a check which evaluates whether the
containerd file is already present in the cluster-template.yaml.
This makes configure_containerd.sh idempotent.

Issue #471

Signed-off-by: Matej Feder <matej.feder@dnation.cloud>

* Apply suggestions from code review

Co-authored-by: Roman Hros <roman.hros@dnation.cloud>
Signed-off-by: Matej Feder <matej.feder@dnation.cloud>

---------

Signed-off-by: Matej Feder <matej.feder@dnation.cloud>
Co-authored-by: Roman Hros <roman.hros@dnation.cloud>