Feat/flavors v3 (#424)
* Create boot volume for kind, default to SSD flavor for capo.

With flavor spec v3, the flavors with disks are recommended only.
So create a boot volume for kind and default to the (newly mandatory)
SSD flavors for the k8s control plane and workers.

Also detect external network and default to it if not set.

* Avoid non-existent flavor on wavestack.

* Error out on conflicting unknown app cred.

* Calculate needed volume size for flavor/image.

Check if flavor has a disk; if not, suggest size for volume
to be created. Checks existence of flavors, optionally image,
and ensures it's large enough.

* Call flavor disk check and report errors.

* Dynamically patch cluster-template with rootVolume.

* Remove empty lines (yamllint).

* Install jq as we need it.

* Avoid changing into not-yet-existing dir.

* Use tf-1.3.6 for github workflow.

* Mention new parameters in documentation.

* Fixup (default) 0-sized volumes.

* Mention support for diskless flavors in the release notes.

* Fix terraform fmt -check errors.

* Fix DISKIZE -> DISKSIZE typo.

Thanks @chess-knight!

Signed-off-by: Kurt Garloff <kurt@garloff.de>
Co-authored-by: Jan Schoone <6106846+jschoone@users.noreply.github.com>

(cherry picked from commit c585b006e18777b895d212c2636f791da9f22a8a)
Signed-off-by: Filip Dobrovolny <dobrovolny.filip@gmail.com>