New config-setting no_proxy
This addresses Issue #418 and introduces a new config variable
name no_proxy. It is trickled down in the same corners where http_proxy
is (when possible). A note to reviewers: When no "no_proxy" variable is
set (from terraforms side). The variable "no_proxy" ends with a trailing
"," (Because i did not spend the effort of adding another case for an empty
"no_proxy" variable in terraform). But this seems to work. Neither could
i notice that a program was confused by the trailing slash nor did any
program treat this as "bypass proxy for all requests".

Signed-off-by: Malte Münch <muench@b1-systems.de>