Fix multiple namespaces. Deploy Secret.
So we deploy helm into a namespace and avoid helm managing the
namespace. This would otherwise result in helm deleting the
old namespace and the CRS when we want to use another namespace.
This would leave CAPI in a really bad state, unable to talk to
the infra, unable to clean up.

Signed-off-by: Kurt Garloff <kurt@garloff.de>